Refactor: alignmentSlice
Move alignment data and selected text segments to same
state slice to tidy up for createLink() action.
