+++ src/app/store.ts
@@ -2,14 +2,14 @@ import { configureStore } from '@reduxjs/toolkit';
 import undoable from 'redux-undo';
 
 import AppReducer from 'state/app.slice';
-import TextSegmentReducer from 'state/textSegment.slice';
+import AlignmentReducer from 'state/alignment.slice';
 import TextSegmentHoverReducer from 'state/textSegmentHover.slice';
 import PolyglotReducer from 'state/polyglot.slice';
 
 export const store = configureStore({
   reducer: {
     app: AppReducer,
-    textSegment: undoable(TextSegmentReducer),
+    alignment: undoable(AlignmentReducer),
     textSegmentHover: TextSegmentHoverReducer,
     polyglot: PolyglotReducer,
   },
